style: move things to node dir
