+++ packages/instrumentation-exception/src/index.ts
@@ -1,2 +1,3 @@
+export { recordException } from './node';
 export * from './instrumentation';
 export * from './types';
