+++ packages/instrumentation-exception/src/instrumentation.ts
@@ -1,36 +1,13 @@
-import type { EventHint } from '@sentry/types';
 import {
   InstrumentationNodeModuleDefinition,
   InstrumentationBase,
 } from '@opentelemetry/instrumentation';
 import { TracerProvider, diag, trace } from '@opentelemetry/api';
-import { getEventProcessor } from '@hyperdx/instrumentation-sentry-node';
 
 import { ExceptionInstrumentationConfig } from './types';
-import { SENTRY_SDK_VERSION, buildEventFromException } from './sentry';
 import { name as PKG_NAME, version as PKG_VERSION } from '../package.json';
-import { onUncaughtExceptionIntegration } from './sentry/integrations/onuncaughtexception';
-import { onUnhandledRejectionIntegration } from './sentry/integrations/onunhandledrejection';
-
-const defaultTracer = trace.getTracer(PKG_NAME, PKG_VERSION);
-const _eventProcessor = getEventProcessor(defaultTracer, SENTRY_SDK_VERSION);
-
-export const recordException = async (e: any, hint?: EventHint) => {
-  try {
-    const event = await buildEventFromException(e, hint);
-    _eventProcessor(
-      event,
-      hint ?? {
-        mechanism: {
-          type: 'generic',
-          handled: false,
-        },
-      },
-    );
-  } catch (err) {
-    diag.error('Failed to capture exception', err);
-  }
-};
+import { onUncaughtExceptionIntegration } from './node/integrations/onuncaughtexception';
+import { onUnhandledRejectionIntegration } from './node/integrations/onunhandledrejection';
 
 /** Exception instrumentation for OpenTelemetry */
 export class ExceptionInstrumentation extends InstrumentationBase {
