+++ packages/instrumentation-exception/src/node/index.ts
@@ -1,10 +1,8 @@
 import type { Client, Event, EventHint, Integration } from '@sentry/types';
 import { SDK_VERSION, inboundFiltersIntegration } from '@sentry/core';
-import {
-  eventFromUnknownInput,
-  timestampInSeconds,
-  uuid4,
-} from '@sentry/utils';
+import { diag, trace } from '@opentelemetry/api';
+import { eventFromUnknownInput } from '@sentry/utils';
+import { getEventProcessor } from '@hyperdx/instrumentation-sentry-node';
 
 import { contextLinesIntegration } from './integrations/contextlines';
 import { defaultStackParser } from './sdk/api';
@@ -13,6 +11,11 @@ import { isCjs } from './utils/commonjs';
 import { localVariablesIntegration } from './integrations/local-variables';
 import { modulesIntegration } from './integrations/modules';
 import { nodeContextIntegration } from './integrations/context';
+import { name as PKG_NAME, version as PKG_VERSION } from '../../package.json';
+
+// TODO: does it make sense to have a default tracer here?
+const defaultTracer = trace.getTracer(PKG_NAME, PKG_VERSION);
+const _eventProcessor = getEventProcessor(defaultTracer, SDK_VERSION);
 
 function getCjsOnlyIntegrations(): Integration[] {
   return isCjs() ? [modulesIntegration()] : [];
@@ -46,9 +49,6 @@ export const applyIntegrations = async (event: Event, hint?: EventHint) => {
   return event;
 };
 
-//
-export const SENTRY_SDK_VERSION = SDK_VERSION;
-
 export const buildEventFromException = async (
   exception: unknown,
   hint?: EventHint,
@@ -60,4 +60,21 @@ export const buildEventFromException = async (
     hint,
   );
   return applyIntegrations(event, hint);
+};
+
+export const recordException = async (e: any, hint?: EventHint) => {
+  try {
+    const event = await buildEventFromException(e, hint);
+    _eventProcessor(
+      event,
+      hint ?? {
+        mechanism: {
+          type: 'generic',
+          handled: false,
+        },
+      },
+    );
+  } catch (err) {
+    diag.error('Failed to capture exception', err);
+  }
 };
