+++ packages/instrumentation-exception/src/node/integrations/local-variables/common.ts
