+++ packages/instrumentation-exception/src/node/integrations/local-variables/index.ts
