+++ packages/instrumentation-exception/src/node/integrations/local-variables/inspector.d.ts
