+++ packages/instrumentation-exception/src/node/integrations/local-variables/local-variables-sync.ts
