+++ packages/instrumentation-exception/src/node/integrations/local-variables/worker.ts
