+++ packages/instrumentation-exception/src/node/integrations/modules.ts
