+++ packages/instrumentation-exception/src/node/integrations/onuncaughtexception.ts
@@ -2,7 +2,7 @@ import { defineIntegration } from '@sentry/core';
 import { diag } from '@opentelemetry/api';
 
 import { logAndExitProcess } from '../utils/errorhandling';
-import { recordException } from '../../instrumentation';
+import { recordException } from '..';
 
 type OnFatalErrorHandler = typeof logAndExitProcess;
