+++ packages/instrumentation-exception/src/node/nodeVersion.ts
