+++ packages/instrumentation-exception/src/node/utils/commonjs.ts
