+++ packages/instrumentation-exception/src/node/utils/module.ts
